# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Java 25 SQLite Template API
  version: 0.1.0
servers:
  - url: /
paths:
  /api/health:
    get:
      responses:
        '200':
          description: Health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/csrf:
    get:
      responses:
        '200':
          description: CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
  /api/auth/login:
    post:
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      security:
        - refreshCookie: []
          csrfHeader: []
      responses:
        '200':
          description: Rotated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
  /api/auth/logout:
    post:
      security:
        - csrfHeader: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
  /api/auth/me:
    get:
      security:
        - accessCookie: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
  /api/protected/profile:
    get:
      security:
        - accessCookie: []
      responses:
        '200':
          description: Protected profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectedProfileResponse'
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status: { type: string, enum: [ok] }
        service: { type: string }
    CsrfResponse:
      type: object
      required: [token]
      properties:
        token: { type: string }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    LogoutResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    ProtectedProfileResponse:
      type: object
      required: [profile]
      properties:
        profile:
          type: object
          required: [email, role]
          properties:
            email: { type: string, format: email }
            role: { type: string, enum: [admin, member] }
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message: { type: string }
        code: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [admin, member] }